ZOOM/requirements.txt
aniso8601==3.0.2
certifi==2018.10.15
chardet==3.0.4
Click==7.0
croniter==0.3.25
Django==2.1.2
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-extensions==2.1.3
django-filter==2.0.0
django-grappelli==2.11.1
django-modeladmin-reorder==0.3.1
django-multiselectfield==0.1.8
django-redis-cache==1.7.1
django-rq==1.2.0
django-test-without-migrations==0.6
djangorestframework==3.8.2
djangorestframework-csv==2.1.0
factory-boy==2.11.1
Faker==0.9.2
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
idna==2.7
LEPL==5.1.3
lxml==4.2.5
mock==2.0.0
numpy==1.15.2
pandas==0.23.4
pbr==4.3.0
promise==2.2.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
redis==2.10.6
requests==2.19.1
rfc6266==0.0.4
rq==0.12.0
rq-scheduler==0.8.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.12
sqlparse==0.2.4
strict-rfc3339==0.7
text-unidecode==1.2
ujson==1.35
unicodecsv==0.14.1
urllib3==1.23
